chore: 테스트 클래스 명 변경
