+++ src/test/java/kitchenpos/product/acceptance/ProductAcceptanceTest.java
@@ -1,69 +1,62 @@
 package kitchenpos.product.acceptance;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.mockito.BDDMockito.*;
 
 import java.math.BigDecimal;
-import java.util.Arrays;
-import java.util.List;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 
-import kitchenpos.application.ProductService;
-import kitchenpos.dao.ProductDao;
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import kitchenpos.AcceptanceTest;
 import kitchenpos.domain.Product;
 
-@ExtendWith(MockitoExtension.class)
-public class ProductAcceptanceTest {
-    @InjectMocks
-    private ProductService productService;
-
-    @Mock
-    private ProductDao productDao;
-
-    private Product product;
-
-    @BeforeEach
-    void setUp() {
-        product = 상품_정보(1L, "제육볶음", 8900);
-    }
+@DisplayName("상품 관련 기능")
+public class ProductAcceptanceTest extends AcceptanceTest {
 
     @DisplayName("상품을 등록한다.")
     @Test
     void create() {
-        // given
-        given(productDao.save(any(Product.class))).willReturn(product);
-        Product expectedProduct = productService.create(product);
+        ExtractableResponse 상품_등록_요청_응답 = 상품_등록_요청("제육볶음", 8900);
 
-        // then
-        assertThat(expectedProduct).isNotNull();
-        assertThat(expectedProduct.getName()).isEqualTo("제육볶음");
-        assertThat(expectedProduct.getPrice()).isEqualTo(BigDecimal.valueOf(8900));
+        상품_등록_요청됨(상품_등록_요청_응답);
     }
 
     @DisplayName("상품 목록을 조회한다.")
     @Test
-    void list(){
-        // given
-        given(productDao.findAll()).willReturn(Arrays.asList(product));
-        List<Product> products = productService.list();
+    void list() {
+        ExtractableResponse 상품_목록_요청_응답 = 상품_목록_요청();
 
-        // then
-        assertThat(products).isNotNull();
-        assertThat(products.size()).isEqualTo(1);
+        상품_목록_요청됨(상품_목록_요청_응답);
     }
 
-    private Product 상품_정보(Long id, String name, int price) {
+    private ExtractableResponse 상품_목록_요청() {
+        return RestAssured
+            .given().log().all()
+            .when().get("/api/products")
+            .then().log().all().extract();
+    }
+
+    private ExtractableResponse 상품_등록_요청(String name, int price) {
         Product product = new Product();
-        product.setId(id);
         product.setName(name);
         product.setPrice(BigDecimal.valueOf(price));
-        return product;
+        return RestAssured
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(product)
+            .when().post("/api/products")
+            .then().log().all().extract();
+    }
+
+    private void 상품_등록_요청됨(ExtractableResponse response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+    }
+
+    private void 상품_목록_요청됨(ExtractableResponse response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
     }
 }
